requirements.txt
# Django y core (usando 4.2 LTS por compatibilidad con celery-beat)
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.5
django-cors-headers==4.3.1

# Base de datos
psycopg2-binary==2.9.9
redis==5.0.1
dj-database-url==2.1.0

# Celery
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.0
flower==2.0.1

# Autenticación y seguridad
cryptography==42.0.2
PyJWT==2.8.0

# Web scraping
selenium==4.17.2
beautifulsoup4==4.12.2
requests==2.31.0
webdriver-manager==4.0.1

# OpenAI and LangChain - latest compatible versions
openai>=1.24.0
langchain>=0.3.0
langchain-openai
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langgraph-supervisor
langsmith
pydantic>=2.7.4

# Vector database and scientific computing
faiss-cpu==1.7.4
numpy>=1.24.0,<2.0.0  # Compatible with FAISS

# Utilidades
python-decouple==3.8
Pillow==10.2.0
python-dateutil==2.8.2

# Security dependencies
bleach==6.1.0
psutil==5.9.8

# Formateadores y validadores
phonenumbers==8.13.27
validators==0.22.0

# Exportación
openpyxl==3.1.2
reportlab==4.0.9

# Logging y monitoreo
sentry-sdk[django]==1.40.6

# Calidad de código
flake8==7.0.0
black==24.1.1
isort==5.13.2

# Production server
gunicorn==21.2.0

# Static files
whitenoise==6.6.0

# Health checks
django-health-check==3.17.0

# Redis cache (Railway compatible)
django-redis==5.4.0

# Rate limiting
django-ratelimit==4.1.0

# Testing (for Railway CI/CD if needed)
pytest==8.0.0
pytest-django==4.8.0

# Debugging tools for production troubleshooting
django-extensions==3.2.3